refactor args helper.
