+++ packages/stitch/src/index.ts
@@ -1,1 +1,2 @@
 export { stitchSchemas } from './stitchSchemas';
+export { forwardArgsToSelectionSet } from './selectionSetArgs';
